# docs/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Online Supermarket System API",
    "description": "RESTful API specification for the Multi-branch Online Supermarket System (.NET 10 Minimal APIs).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5072",
      "description": "Local Development Server"
    },
    {
      "url": "http://localhost:8080",
      "description": "Docker Container API Gateway"
    }
  ],
  "paths": {
...
    "/api/users/me": {
      "put": {
        "tags": ["Users"],
        "summary": "Update user profile",
        "operationId": "UpdateProfile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/users/me/password": {
      "put": {
        "tags": ["Users"],
        "summary": "Change password",
        "operationId": "ChangePassword",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Current password incorrect or validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": ["Admin/Users"],
        "summary": "List users (Admin)",
        "operationId": "ListUsers",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsersDto"
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden - Admin role required" }
        }
      }
    },
    "/api/admin/users/{id}/status": {
      "put": {
        "tags": ["Admin/Users"],
        "summary": "Update user status (Lock/Unlock/Disable)",
        "operationId": "UpdateUserStatus",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User status updated" },
          "400": { "description": "Invalid status value" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "User not found" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateProfileRequest": {
        "type": "object",
        "required": ["fullName"],
        "properties": {
          "fullName": { "type": "string" },
          "phone": { "type": "string", "nullable": true }
        }
      },
      "ChangePasswordRequest": {
        "type": "object",
        "required": ["currentPassword", "newPassword"],
        "properties": {
          "currentPassword": { "type": "string" },
          "newPassword": { "type": "string", "minLength": 6 }
        }
      },
      "UpdateUserStatusRequest": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "type": "string", "enum": ["Active", "Locked", "Disabled"] }
        }
      },
      "UserSummaryDto": {
        "type": "object",
        "required": ["id", "email", "fullName", "role", "status", "createdAtUtc", "updatedAtUtc"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "email": { "type": "string" },
          "fullName": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "role": { "type": "string" },
          "status": { "type": "string" },
          "createdAtUtc": { "type": "string", "format": "date-time" },
          "updatedAtUtc": { "type": "string", "format": "date-time" }
        }
      },
      "PaginatedUsersDto": {
        "type": "object",
        "required": ["page", "pageSize", "totalCount", "users"],
        "properties": {
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "totalCount": { "type": "integer" },
          "users": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/UserSummaryDto" }
          }
        }
      },
...
    }
  }
}